Time: 0 ms (100.00%), Space: 8.3 MB (72.00%) - LeetHub

+++ 0050-powx-n/0050-powx-n.cpp
@@ -12,7 +12,8 @@ public:
             ans = ans*x; nn = nn-1;
         }
        }
-       if(n<0) ans = (double)(1.0)/(double)(ans);
+       if(n<0) ans = (double)(1.0)/ans;
        return ans;
+    
     }
 };
